+++ .gitignore
@@ -1,9 +1,24 @@
-# Backend
-backend/venv
-backend/LIC_Connect/__pycache__
-backend/students/__pycache__
-backend/students/migrations
+# Logs
+logs
+*.log
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+pnpm-debug.log*
+lerna-debug.log*
 
-# Frontend
+node_modules
+dist
+dist-ssr
+*.local
 
-frontend/node_modules
+# Editor directories and files
+.vscode/*
+!.vscode/extensions.json
+.idea
+.DS_Store
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
+*.sw?
