+++ LIC_Connect/manage.py
