--- frontend/src/Components/Logout.jsx
@@ -1,20 +1,0 @@
-import axios from "axios";
-import { useNavigate } from "react-router-dom";
-
-const Logout = () => {
-  const navigate = useNavigate();
-
-  const handleLogout = async () => {
-    try {
-      await axios.post("/api/logout/"); // Adjust the URL to your logout endpoint
-      // On success, navigate to the login page or home page
-      navigate("/login");
-    } catch (error) {
-      console.error("Logout failed:", error);
-    }
-  };
-
-  return <button onClick={handleLogout}>Logout</button>;
-};
-
-export default Logout;
