--- frontend/src/Login/Login.css
@@ -1,14 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  font-family: 'Arial', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif
-}
-
-.loginHeader {
-  margin-bottom: 20px;
-  text-align: center;
-  font-family: 'Poppins', sans-serif;
-
-
-}
