+++ package.json
@@ -16,7 +16,6 @@
     "@mui/icons-material": "^6.0.2",
     "@mui/joy": "^5.0.0-beta.48",
     "@mui/material": "^6.0.2",
-    "axios": "^1.7.7",
     "lic-reactver": "file:",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
