+++ src/Settings/Settings.css
@@ -2,6 +2,4 @@
   background-color: white;
   border-radius: 5px;
   padding: 0 25px 25px;
-}
-
-;
+};
